calender cell buton added and query params added to the url
